+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#3](https://github.com/LaxarJS/grunt-laxar-compass/issues/3): the target laxar-compass can now rebuild all artifacts in batch
+    + NEW FEATURE: see ticket for details
 - [#4](https://github.com/LaxarJS/grunt-laxar-compass/issues/4): allow to rebuild individual artifacts
     + NEW FEATURE: see ticket for details
 - [#2](https://github.com/LaxarJS/grunt-laxar-compass/issues/2): watch layout SCSS, not just controls/widgets/themes